Finish GET project endpoint

+++ domain/src/main/java/onlydust/com/marketplace/api/domain/model/UserProfile.java
@@ -12,7 +12,7 @@ import java.util.*;
 @Builder
 public class UserProfile {
     UUID id;
-    Integer githubId;
+    Long githubId;
     String bio;
     String avatarUrl;
     String login;
